refactor(Line): Ladder 이름 변경
- 이전 변경에 따른 Line 내부 dto 이름 및 method 이름 변경
