+++ src/main/java/ladder/domain/dto/StepStatusDto.java
@@ -1,7 +1,7 @@
 package ladder.domain.dto;
 
 import java.util.List;
-import ladder.domain.Step;
+import ladder.domain.ladder.Step;
 
 public record StepStatusDto(List<Step> builtStep) {
 }
