+++ src/main/java/ladder/domain/ladder/Line.java
@@ -1,11 +1,10 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.ArrayList;
 import java.util.List;
-import ladder.domain.dto.BuiltLadderDto;
 import ladder.domain.dto.StepStatusDto;
 
-public class Ladder {
+public class Line {
 
     private static final int STEP_START_NUMBER = 0;
 
@@ -13,15 +12,11 @@ public class Ladder {
 
     private final List<Step> steps;
 
-    public Ladder(final int personCount) {
+    public Line(final int personCount) {
         this.steps = makeSteps(personCount);
     }
 
-    public BuiltLadderDto getSteps() {
-        return BuiltLadderDto.of(steps);
-    }
-
-    public StepStatusDto getSteps2() {
+    public StepStatusDto getSteps() {
         return new StepStatusDto(steps);
     }
 
@@ -32,6 +27,7 @@ public class Ladder {
     public boolean hasStepDuplicated(int currentPosition) {
         if (currentPosition > STEP_START_NUMBER) {
             int stepBeforePosition = currentPosition - 1;
+
             Step step = steps.get(stepBeforePosition);
             return step.getBuildStatus();
         }
