+++ src/test/java/ladder/domain/LineTest.java
@@ -2,7 +2,8 @@ package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import ladder.domain.dto.BuiltLadderDto;
+import ladder.domain.dto.StepStatusDto;
+import ladder.domain.ladder.Line;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -10,17 +11,17 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class LadderTest {
+public class LineTest {
 
     private static final int FIRST_STEP_INDEX = 0;
     private static final int NEXT_STEP_INDEX = 1;
     private static final int NOT_BUILD_STEP_INDEX = 3;
 
-    private Ladder ladder;
+    private Line line;
 
     @BeforeEach
     void setUp() {
-        ladder = new Ladder(5);
+        line = new Line(5);
     }
 
     @Nested
@@ -30,9 +31,9 @@ public class LadderTest {
         @Test
         @DisplayName("이전 포지션에 스텝이 존재하면 True를 반환한다.")
         void hasDuplicatedStep() {
-            ladder.buildSteps(FIRST_STEP_INDEX);
+            line.buildSteps(FIRST_STEP_INDEX);
 
-            Boolean hasDuplicateStep = ladder.hasStepDuplicated(NEXT_STEP_INDEX);
+            Boolean hasDuplicateStep = line.hasStepDuplicated(NEXT_STEP_INDEX);
 
             assertThat(hasDuplicateStep).isTrue();
         }
@@ -40,9 +41,9 @@ public class LadderTest {
         @Test
         @DisplayName("이전 포지션에 스텝이 존재하지 않으면 False를 반환한다.")
         void hasNotDuplicatedStep() {
-            ladder.buildSteps(FIRST_STEP_INDEX);
+            line.buildSteps(FIRST_STEP_INDEX);
 
-            Boolean hasDuplicateStep = ladder.hasStepDuplicated(NOT_BUILD_STEP_INDEX);
+            Boolean hasDuplicateStep = line.hasStepDuplicated(NOT_BUILD_STEP_INDEX);
 
             assertThat(hasDuplicateStep).isFalse();
         }
@@ -56,10 +57,10 @@ public class LadderTest {
         @ValueSource(ints = {2, 3, 4})
         @DisplayName("스텝을 건설한 구역은 True를 반환한다.")
         void successfullyMakeStepTest(int currentSector) {
-            ladder.buildSteps(currentSector);
+            line.buildSteps(currentSector);
 
-            BuiltLadderDto builtLadderDto = ladder.getSteps();
-            assertThat(builtLadderDto.builtLadder().get(currentSector)).isEqualTo("-----");
+            StepStatusDto builtLadderDto = line.getSteps();
+            assertThat(builtLadderDto.builtStep().get(currentSector)).isEqualTo("-----");
         }
     }
 }
